requirements.txt
llama-cpp-python==0.1.83
huggingface-hub==0.14.1
gradio==3.32.0